# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AI Chatbot Law API",
    "version": "1.0.0",
    "description": "API documentation for the AI Chatbot for Education Laws."
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Local server"
    }
  ],
  "paths": {
...
    "/legal-docs": {
      "get": {
        "summary": "[USER/ADMIN] Get all legal documents",
        "description": "User và admin đều có thể xem danh sách văn bản pháp luật.",
        "tags": ["LegalDocument"],
        "responses": { "200": { "description": "List of legal documents" } }
      },
      "post": {
        "summary": "[ADMIN] Create a new legal document",
        "description": "Chỉ admin mới được tạo mới văn bản pháp luật.",
        "tags": ["LegalDocument"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "document_name": { "type": "string" },
                  "document_type": { "type": "string" },
                  "document_date_issue": { "type": "string", "format": "date" },
                  "document_signee": { "type": "string" }
                },
                "required": [
                  "document_name",
                  "document_type",
                  "document_date_issue"
                ]
              }
            }
          }
        },
        "responses": { "201": { "description": "Legal document created" } }
      }
    },
...
    "/legal-clauses": {
      "get": {
        "summary": "[USER/ADMIN] Get all legal clauses",
        "description": "User và admin đều có thể xem danh sách điều khoản.",
        "tags": ["LegalClause"],
        "responses": { "200": { "description": "List of legal clauses" } }
      },
      "post": {
        "summary": "[ADMIN] Create a new legal clause",
        "description": "Chỉ admin mới được tạo mới điều khoản.",
        "tags": ["LegalClause"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "clause_number": { "type": "string" },
                  "clause_content": { "type": "string" },
                  "embedding": {
                    "type": "array",
                    "items": { "type": "number" }
                  },
                  "document_id": { "type": "string" }
                },
                "required": ["clause_number", "clause_content", "document_id"]
              }
            }
          }
        },
        "responses": { "201": { "description": "Legal clause created" } }
      }
    },
...
    "/notifications": {
      "get": {
        "summary": "[USER/ADMIN] Get all notifications",
        "description": "User và admin đều có thể xem danh sách thông báo.",
        "tags": ["Notification"],
        "responses": { "200": { "description": "List of notifications" } }
      },
      "post": {
        "summary": "[ADMIN] Create a new notification",
        "description": "Chỉ admin mới được tạo mới thông báo.",
        "tags": ["Notification"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" },
                  "content": { "type": "string" },
                  "isRead": { "type": "boolean" }
                },
                "required": ["title", "content"]
              }
            }
          }
        },
        "responses": { "201": { "description": "Notification created" } }
      }
    },
...
    "/questions": {
      "get": {
        "summary": "[ADMIN] Get all questions",
        "description": "Chỉ admin mới được xem danh sách tất cả câu hỏi.",
        "tags": ["Question"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of questions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Question" }
                }
              }
            }
          },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" }
        }
      }
    },
    "/feedbacks": {
      "post": {
        "summary": "[USER] Gửi feedback về hệ thống",
        "description": "User gửi góp ý/phản hồi về hệ thống.",
        "tags": ["Feedback"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "user_id": { "type": "string" },
                  "content": { "type": "string" }
                },
                "required": ["user_id", "content"]
              }
            }
          }
        },
        "responses": { "201": { "description": "Feedback created" } }
      },
      "get": {
        "summary": "[ADMIN] Xem tất cả feedback",
        "description": "Admin xem danh sách góp ý/phản hồi của user.",
        "tags": ["Feedback"],
        "responses": { "200": { "description": "List of feedbacks" } }
      }
    },
...
    "/chat-room/create": {
      "post": {
        "summary": "Tạo phòng chat mới",
        "description": "Tạo phòng chat mới cho user với chat_id tự động sinh.",
        "tags": ["ChatRoom"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "account_id": {
                    "type": "string",
                    "description": "ID của user tạo phòng chat"
                  },
                  "room_name": {
                    "type": "string",
                    "description": "Tên phòng chat (optional)"
                  }
                },
                "required": ["account_id"]
              },
              "example": {
                "account_id": "user_id",
                "room_name": "Chủ đề Luật Lao Động"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Phòng chat được tạo thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "chat_id": { "type": "string" },
                    "room": { "$ref": "#/components/schemas/ChatRoom" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Dữ liệu đầu vào không hợp lệ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/chat-room/list": {
      "get": {
        "summary": "Lấy danh sách phòng chat của user",
        "description": "Lấy tất cả phòng chat của một user theo account_id.",
        "tags": ["ChatRoom"],
        "parameters": [
          {
            "name": "account_id",
            "in": "query",
            "required": true,
            "description": "ID của user để lấy danh sách phòng chat",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Danh sách phòng chat",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/ChatRoom" }
                }
              }
            }
          },
          "400": {
            "description": "Thiếu account_id",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/chat-room/message/send": {
      "post": {
        "summary": "Gửi tin nhắn vào phòng chat (tự động sinh answer với AI)",
        "description": "Gửi câu hỏi, hệ thống sẽ tìm kiếm điều khoản liên quan và tạo câu trả lời bằng AI (Gemini).",
        "tags": ["ChatRoom"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "question_content": {
                    "type": "string",
                    "description": "Nội dung câu hỏi về pháp luật"
                  },
                  "account_id": {
                    "type": "string",
                    "description": "ID của user gửi câu hỏi"
                  },
                  "chat_id": {
                    "type": "string",
                    "description": "ID phòng chat (optional, sẽ tự tạo nếu không có)"
                  }
                },
                "required": ["question_content", "account_id"]
              },
              "example": {
                "question_content": "Quy định về nghỉ thai sản theo luật lao động?",
                "account_id": "user_id",
                "chat_id": "chat_id_phong"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tin nhắn và câu trả lời được tạo thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "question": {
                          "type": "object",
                          "properties": {
                            "id": { "type": "string" },
                            "content": { "type": "string" },
                            "date": { "type": "string", "format": "date-time" }
                          }
                        },
                        "answer": {
                          "type": "object",
                          "properties": {
                            "id": { "type": "string" },
                            "ai_response": {
                              "type": "string",
                              "description": "Câu trả lời ngắn gọn của AI"
                            },
                            "full_content": {
                              "type": "string",
                              "description": "Câu trả lời đầy đủ bao gồm cả điều khoản"
                            },
                            "date": { "type": "string", "format": "date-time" }
                          }
                        },
                        "related_clauses": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "stt": { "type": "integer" },
                              "dieu": { "type": "string" },
                              "noi_dung": { "type": "string" },
                              "van_ban": { "type": "string" },
                              "document_id": { "type": "string" }
                            }
                          }
                        },
                        "chat_id": { "type": "string" },
                        "metadata": {
                          "type": "object",
                          "properties": {
                            "keyword_used": { "type": "string" },
                            "search_terms": {
                              "type": "array",
                              "items": { "type": "string" }
                            },
                            "clauses_found": { "type": "integer" },
                            "documents_involved": {
                              "type": "array",
                              "items": { "type": "string" }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Dữ liệu đầu vào không hợp lệ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "error": { "type": "string" },
                    "details": { "type": "string" }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Duplicate entry",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "error": { "type": "string" }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Lỗi server",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "error": { "type": "string" },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/chat-room/{chat_id}/messages": {
      "get": {
        "summary": "Lấy lịch sử chat của phòng",
        "description": "Lấy tất cả tin nhắn (question-answer pairs) trong một phòng chat.",
        "tags": ["ChatRoom"],
        "parameters": [
          {
            "name": "chat_id",
            "in": "path",
            "required": true,
            "description": "ID của phòng chat",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Lịch sử chat của phòng",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/ChatMessage" }
                }
              }
            }
          },
          "500": {
            "description": "Lỗi server",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Nhập JWT token vào đây. Ví dụ: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
      }
    },
    "schemas": {
...
      "Question": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "question_id": { "type": "string" },
          "question_content": { "type": "string" },
          "question_date": { "type": "string", "format": "date-time" },
          "account_id": { "type": "string" },
          "chat_id": { "type": "string" }
        }
      },
...
      "ChatRoom": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "chat_id": { "type": "string" },
          "account_id": { "type": "string" },
          "room_name": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "ChatMessage": {
        "type": "object",
        "properties": {
          "question": { "type": "string" },
          "answer": { "type": "string" },
          "question_date": { "type": "string", "format": "date-time" },
          "answer_date": { "type": "string", "format": "date-time" },
          "chat_id": { "type": "string" }
        }
      },
...
    }
  }
}